refactored structs
